Implement the `build_info` functions for the C and WebAssembly frontends
This gives C and WebAssembly users information about how
Babycat was compiled.
