+++ src/frontends/c/mod.rs
@@ -1,0 +1,9 @@
+mod build_info;
+mod error;
+mod waveform;
+mod waveform_result;
+
+pub use build_info::*;
+pub use error::*;
+pub use waveform::*;
+pub use waveform_result::*;
