+++ src/frontends/wasm/mod.rs
@@ -1,0 +1,7 @@
+mod build_info;
+mod error;
+mod waveform;
+
+pub use build_info::*;
+pub use error::*;
+pub use waveform::*;
